refactor : ViewController 그룹 나눔
